Added recent work changed, plot and sorting utils

+++ source/utils/triplets_gen.py
@@ -10,20 +10,23 @@ import cv2
 import os
 import numpy as np
 
-src_path_A = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/sorted_whole/plants_RGB_A'
-dst_path_A = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/triplets/RGB_A'
-src_path_B = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/sorted_whole/plants_RGB_B'
-dst_path_B = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/triplets/RGB_B'
-src_path_C = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/sorted_whole/plants_RGB_C'
-dst_path_C = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/triplets/RGB_C'
-src_path_D = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/sorted_whole/plants_RGB_D'
-dst_path_D = 'D:/Users Data/ItamarGIP/Desktop/Itamar/data/seg_data/triplets/RGB_D'
 
-# src_paths = [src_path_A, src_path_B, src_path_C, src_path_D]
-# dst_paths = [dst_path_A, dst_path_B, dst_path_C, dst_path_D]
+src_path_A = '/home/itamargoz/data/segmented/SEG_RGB_A'
+src_path_B = '/home/itamargoz/data/segmented/SEG_RGB_B'
+src_path_C = '/home/itamargoz/data/segmented/SEG_RGB_C'
+src_path_D = '/home/itamargoz/data/segmented/SEG_RGB_D'
 
-src_paths = [src_path_B, src_path_C, src_path_D]
-dst_paths = [dst_path_B, dst_path_C, dst_path_D]
+
+dst_path_A = '/home/itamargoz/data/segmented/triplets/RGB_A'
+dst_path_B = '/home/itamargoz/data/segmented/triplets/RGB_B'
+dst_path_C = '/home/itamargoz/data/segmented/triplets/RGB_C'
+dst_path_D = '/home/itamargoz/data/segmented/triplets/RGB_D'
+
+src_paths = [src_path_A] #, src_path_B, src_path_C, src_path_D]
+dst_paths = [dst_path_A] #, dst_path_B, dst_path_C, dst_path_D]
+
+# src_paths = [src_path_B, src_path_C, src_path_D]
+# dst_paths = [dst_path_B, dst_path_C, dst_path_D]
 
 
 def crate_target_name(src_a, src_b, src_c):
@@ -43,17 +46,23 @@ def join_pictures(src_path, dst_path, images):
     print(("========== Join picutres =========="))
     dim = (336, 252)
     for i in range(2, len(images)):
-        image1 = cv2.imread(src_path + "/" + images[i - 2])
-        image2 = cv2.imread(src_path + "/" + images[i - 1])
-        image3 = cv2.imread(src_path + "/" + images[i])
-        image1 = cv2.resize(image1, dim)
-        image2 = cv2.resize(image2, dim)
-        image3 = cv2.resize(image3, dim)
-
-        target_img = cv2.hconcat([image1, image2, image3])
         new_name = crate_target_name(images[i - 2], images[i - 1], images[i])
-        write_status = cv2.imwrite(dst_path + "/" + new_name, target_img)
-        print("saving ", dst_path + "/" + new_name, " status = ", str(write_status))
+        if os.path.exists(dst_path + "/" + new_name):
+            print(dst_path + "/" + new_name," exist! skipping")
+            next
+        try:
+            image1 = cv2.imread(src_path + "/" + images[i - 2])
+            image2 = cv2.imread(src_path + "/" + images[i - 1])
+            image3 = cv2.imread(src_path + "/" + images[i])
+            image1 = cv2.resize(image1, dim)
+            image2 = cv2.resize(image2, dim)
+            image3 = cv2.resize(image3, dim)
+
+            target_img = cv2.hconcat([image1, image2, image3])
+            write_status = cv2.imwrite(dst_path + "/" + new_name, target_img)
+            print("saving ", dst_path + "/" + new_name, " status = ", str(write_status))
+        except:
+            print(dst_path + "/" + new_name," ERROR! skipping")
 
 
 def organize_and_execute_join(src_path, dst_path):
